refactor: use better expo route structure
