+++ 3o14/app/(auth)/_layout.tsx
@@ -1,0 +1,17 @@
+import { Stack } from 'expo-router';
+import { useTheme } from '@/hooks/useTheme';
+
+export default function AuthLayout() {
+  const theme = useTheme();
+
+  return (
+    <Stack
+      screenOptions={{
+        headerShown: false,
+        contentStyle: {
+          backgroundColor: theme.colors.background,
+        },
+      }}
+    />
+  );
+}
